autogen_llm_assist/requirements.txt
autogen
pytesseract
Pillow
pyautogui
openai
